models/cgdf_baseline/requirements.txt
tqdm #==4.64.1
configargparse #==1.5.3
h5py #==3.7.0
pyglet #==1.5.27
scikit-learn #==1.0.2
trimesh #==3.14
matplotlib
tensorboard
natsort #==8.2.0
torchinfo
icecream
pykdtree
open3d
wandb